Fix {black,lint} environment re-creation on tox4
I've found that the existing tox configuration, which has the black and
lint environments sharing the same `envdir`, works fine on tox3 but on
tox4 results in the environments being re-created every time one toggles
between them.

There is a suggested solution [on StackOverflow](https://stackoverflow.com/a/64060423),
but it seems not to work as expected on tox4, so I've opened an issue there:
https://github.com/tox-dev/tox/issues/2362

I'll plan to merge this PR once the above issue with tox is resolved.