Remove unnecessary explicit `{toxinidir}` in `tox.ini`
It's documented that the `deps` command runs from within this
directory anyway, plus, we should really be saying
`{toxinidir}{/}requirements-dev.txt` so that the directory
separator used is also platform independent.  But it's simpler
just to shorten the line and match the style of the other
existing `deps` lines in this file.