Determine version from `setuptools_scm`, not `VERSION.txt` (#54)
* Determine version from `setuptools_scm`, not `VERSION.txt`

Closes #49

This change determines the version automatically from the most recent
git tag, if any, rather than requiring us to keep `VERSION.txt` in
sync with the latest tag.  Right now, there is no version tag in
the history of this repo, so the `release` in `docs/conf.py` is
(just prior to the current commit) `0.1.dev68+g83e72bd.d20220423`,
as there have been 68 commits to this repository (and no releases
during that time).  The `version` in `docs/conf.py` is then `0.1`,
and will update if/when later versions are tagged.

I left the definition of `_rootdir` in `docs/conf.py`, as I've
found it useful in other contexts in this file in the other
prototypes I've worked on.  But right now it's not directly used.

* Set `__version__` using `setuptools_scm`

* Mark `PackageNotFoundError` as not covered

For both branches of this exception catching to be covered,
we'd need the package to be installed and not installed
simultaneously.

* Move importlib_metadata requirement

We'll put it before qiskit, since it is a baseline dependency